+++ main.py
@@ -2,44 +2,36 @@
 Starting service of recomendation via FastAPI
 """
 
+import asyncio
 import os
+from pathlib import Path
 from typing import List
 
-import pandas as pd
-import numpy as np
-import uvicorn
 import dill
+import pandas as pd
+import uvicorn
+from dotenv import load_dotenv
 from fastapi import Depends, FastAPI, HTTPException, Security, status
-from fastapi.security.api_key import APIKey, APIKeyHeader, APIKeyQuery
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
+from fastapi.security.api_key import APIKey, APIKeyHeader, APIKeyQuery
 from fastapi.testclient import TestClient
 from pydantic import BaseModel
 from rectools import Columns
 
-from models.models import popoular_number_of_items_days, full_reco_items_list
+from models.models import knn_make_predict, popular_number_of_items_days
 
-API_KEY = "reco_mts_best"
+load_dotenv()
+
+API_KEY = os.getenv("API-KEY")
+BASE_DIR = Path(__file__).parent
+MODELS_DIR = BASE_DIR / "models"
+DATA_DIR = BASE_DIR / "data_original"
 MODEL_LIST = ["pop14d", "user_knn_v1"]
 
 
 class RecoResponse(BaseModel):
     user_id: int
     items: List[int]
-
-
-# Load table of interactions and lists of popular items for last 14 days
-df_inter = pd.read_csv(
-    "./data_original/interactions.csv", parse_dates=["last_watch_dt"]
-)
-df_inter = df_inter.rename(
-    columns={"last_watch_dt": Columns.Datetime, "total_dur": Columns.Weight}
-)
-recolist_pop14d = popoular_number_of_items_days(df_inter)
-recolist_for_cold_users = recolist_pop14d
-
-# Load fitted model for online predictions
-with open("./models/user_knn/model_user_knn.dill", "rb") as f:
-    model_user_knn = dill.load(f)
 
 
 api_key_query = APIKeyQuery(name="api_key", auto_error=False)
@@ -68,13 +60,46 @@ app = FastAPI()
 client = TestClient(app)
 
 
+# Load table of interactions and lists of popular items for last 14 days
+df_inter = pd.read_csv(DATA_DIR / "interactions.csv", parse_dates=["last_watch_dt"])
+df_inter = df_inter.rename(
+    columns={"last_watch_dt": Columns.Datetime, "total_dur": Columns.Weight}
+)
+recolist_pop14d = popular_number_of_items_days(df_inter)
+recolist_for_cold_users = recolist_pop14d
+
+
+class CustomUnpickler(dill.Unpickler):
+    """
+    Это специальный класс для того, чтобы наследовать
+    все зависимости обученной модели.
+    Фишка вся в том, что без такой загрузки uvicorn/gunicorn
+    не будет работать. Можно запустить python main.py, но так
+    работает только с 1 воркером, при большем кол-ве воркеров, а также
+    при запуске с командной строки uvicorn/gunicorn теряются зависимости.
+    Чтобы найти причину и эти 7 строчек ушла почти неделя.
+    Запомнить раз и навсегда!!!
+    """
+
+    def find_class(self, module, name):
+        if name == "my_UserKnn":
+            from models.train_knn import my_UserKnn
+
+            return my_UserKnn
+        return super().find_class(module, name)
+
+
+# Load fitted base model
+knn_model = CustomUnpickler(open(MODELS_DIR / "full_model_user_knn.dill", "rb")).load()
+
+
 @app.get("/health")
 async def root():
     return "Im still alive"
 
 
 @app.get(path="/reco/{model_name}/{user_id}", response_model=RecoResponse)
-async def get_reco(
+def get_reco(
     model_name: str, user_id: int, api_key: APIKey = Depends(get_api_key)
 ) -> RecoResponse:
     if model_name not in MODEL_LIST:
@@ -86,12 +111,11 @@ async def get_reco(
             reco_list = recolist_pop14d
         elif model_name == "user_knn_v1":
             if user_id in df_inter["user_id"].unique():
-                print()
                 # In case for online predictions
                 one_user_df = df_inter[df_inter["user_id"] == user_id][
                     ["user_id", "item_id"]
                 ]
-                reco_list = model_user_knn.predict(one_user_df, 10)
+                reco_list = knn_make_predict(knn_model, one_user_df, recolist_pop14d)
             else:
                 reco_list = recolist_for_cold_users
 
@@ -100,10 +124,14 @@ async def get_reco(
     return reco
 
 
-if __name__ == "__main__":
-
-    host = os.getenv("HOST", "192.168.89.10")  # Вариант для сервера Кирилла
-    # host = os.getenv("HOST", "127.0.0.1")  # Если у себя тестить
+async def main():
+    # host = os.getenv("HOST", "192.168.89.10")  # Вариант для сервера Кирилла
+    host = os.getenv("HOST", "127.0.0.1")  # Если у себя тестить
     port = int(os.getenv("PORT", "8080"))
+    config = uvicorn.Config("main:app", host=host, port=port, workers=4)
+    server = uvicorn.Server(config)
+    await server.serve()
 
-    uvicorn.run(app, host=host, port=port)
+
+if __name__ == "__main__":
+    asyncio.run(main())
