+++ models/models.py
@@ -1,19 +1,19 @@
 """
 Reco Models
 """
-from collections import Counter
-from typing import Dict
+from pathlib import Path
 
 import numpy as np
 import pandas as pd
-import scipy as sp
-import dill
-from implicit.nearest_neighbours import ItemItemRecommender, CosineRecommender
-from rectools import Columns
+
+CURRENT_DIR = Path(__file__).parent
+BASE_DIR = CURRENT_DIR.parents[0]
+MODELS_DIR = CURRENT_DIR
+DATA_DIR = BASE_DIR / "data_original"
 
 
 # For HW1********************************************************************
-def popoular_number_of_items_days(
+def popular_number_of_items_days(
     df: pd.DataFrame, k: int = 10, days: int = 14, all_time: bool = False
 ) -> list:
     """
@@ -33,16 +33,6 @@ def popoular_number_of_items_days(
 
 
 # For HW3********************************************************************
-# Load table of interactions and lists of popular items for last 14 days
-df_inter = pd.read_csv(
-    "./data_original/interactions.csv", parse_dates=["last_watch_dt"]
-)
-df_inter = df_inter.rename(
-    columns={"last_watch_dt": Columns.Datetime, "total_dur": Columns.Weight}
-)
-list_pop_items_14d = popoular_number_of_items_days(df_inter, days=14)
-
-
 def full_reco_items_list(
     arr_reco_after_model: np.array, pop_array: np.array, number: int
 ) -> list:
@@ -79,158 +69,25 @@ def full_reco_items_list(
             return del_repeat_items(full_arr_mix_pop)
 
     full_arr_mix_pop = np.array([])
-    full_arr_mix_pop = np.concatenate(
-        (arr_reco_after_model, pop_array[:number])
-    )
+    full_arr_mix_pop = np.concatenate((arr_reco_after_model, pop_array[:number]))
     full_arr_mix_pop = del_repeat_items(full_arr_mix_pop)
 
     return list(full_arr_mix_pop)
 
 
-class my_UserKnn:
-    """Class for fit-perdict UserKNN model
-    based on ItemKNN model from implicit.nearest_neighbours
+def knn_make_predict(model, data: pd.DataFrame, list_pop_items: list) -> list:
+    """
+    Gets predict after bas-model and add lacking items(to 10)
+    for each user if it needed
     """
 
-    def __init__(
-        self, model: ItemItemRecommender, list_pop_items: list, N_users: int = 50
-    ):
-        self.N_users = N_users
-        self.model = model
-        self.list_pop_items_14d = list_pop_items
-        self.is_fitted = False
+    predict = model.predict(data, 10)
+    predict = predict["item_id"].values
+    final_reco = full_reco_items_list(
+        predict, np.array(list_pop_items), (10 - predict.size)
+    )
 
-    def get_mappings(self, train):
-        self.users_inv_mapping = dict(enumerate(train["user_id"].unique()))
-        self.users_mapping = {v: k for k, v in self.users_inv_mapping.items()}
-
-        self.items_inv_mapping = dict(enumerate(train["item_id"].unique()))
-        self.items_mapping = {v: k for k, v in self.items_inv_mapping.items()}
-
-    def get_matrix(
-        self,
-        df: pd.DataFrame,
-        user_col: str = "user_id",
-        item_col: str = "item_id",
-        weight_col: str = None,
-        users_mapping: Dict[int, int] = None,
-        items_mapping: Dict[int, int] = None,
-    ):
-
-        if weight_col:
-            weights = df[weight_col].astype(np.float32)
-        else:
-            weights = np.ones(len(df), dtype=np.float32)
-
-        interaction_matrix = sp.sparse.coo_matrix(
-            (
-                weights,
-                (
-                    df[user_col].map(self.users_mapping.get),
-                    df[item_col].map(self.items_mapping.get),
-                ),
-            )
-        )
-
-        self.watched = df.groupby(user_col).agg({item_col: list})
-        return interaction_matrix
-
-    def idf(self, n: int, x: float):
-        return np.log((1 + n) / (1 + x) + 1)
-
-    def _count_item_idf(self, df: pd.DataFrame):
-        item_cnt = Counter(df["item_id"].values)
-        item_idf = pd.DataFrame.from_dict(
-            item_cnt, orient="index", columns=["doc_freq"]
-        ).reset_index()
-        item_idf["idf"] = item_idf["doc_freq"].apply(lambda x: self.idf(self.n, x))
-        self.item_idf = item_idf
-
-    def fit(self, train: pd.DataFrame):
-        self.user_knn = self.model
-        self.get_mappings(train)
-        self.weights_matrix = self.get_matrix(
-            train, users_mapping=self.users_mapping, items_mapping=self.items_mapping
-        )
-
-        self.n = train.shape[0]
-        self._count_item_idf(train)
-
-        self.user_knn.fit(self.weights_matrix)
-        self.is_fitted = True
-
-    def _generate_recs_mapper(
-        self,
-        model: ItemItemRecommender,
-        user_mapping: Dict[int, int],
-        user_inv_mapping: Dict[int, int],
-        N: int,
-    ):
-        def _recs_mapper(user):
-            user_id = user_mapping[user]
-            recs = model.similar_items(user_id, N=N)
-            return [user_inv_mapping[user] for user, _ in recs], [
-                sim for _, sim in recs
-            ]
-
-        return _recs_mapper
-
-    def predict(self, test: pd.DataFrame, N_recs: int = 10) -> list:
-
-        if not self.is_fitted:
-            raise ValueError("Please call fit before predict")
-
-        mapper = self._generate_recs_mapper(
-            model=self.user_knn,
-            user_mapping=self.users_mapping,
-            user_inv_mapping=self.users_inv_mapping,
-            N=self.N_users,
-        )
-
-        recs = pd.DataFrame({"user_id": test["user_id"].unique()})
-        recs["sim_user_id"], recs["sim"] = zip(*recs["user_id"].map(mapper))
-        recs = recs.set_index("user_id").apply(pd.Series.explode).reset_index()
-
-        recs = (
-            recs[~(recs["user_id"] == recs["sim_user_id"])]
-            .merge(
-                self.watched, left_on=["sim_user_id"], right_on=["user_id"], how="left"
-            )
-            .explode("item_id")
-            .sort_values(["user_id", "sim"], ascending=False)
-            .drop_duplicates(["user_id", "item_id"], keep="first")
-            .merge(self.item_idf, left_on="item_id", right_on="index", how="left")
-        )
-
-        recs["score"] = recs["sim"] * recs["idf"]
-        recs = recs.sort_values(["user_id", "score"], ascending=False)
-        recs["rank"] = recs.groupby("user_id").cumcount() + 1
-        recs = recs[recs["rank"] <= N_recs]
-
-        # Transforming table and add lacking items(to 10) to each user if it is needed
-        final_reco = recs.groupby("user_id").agg({"item_id": list})
-        final_reco["item_id"] = final_reco.loc[:, "item_id"].apply(
-            lambda x: full_reco_items_list(
-                np.array(x), np.array(self.list_pop_items_14d), (10 - len(x))
-            )
-        )
-
-        reco_for_user = final_reco.loc[test["user_id"].unique(), "item_id"]
-
-        return list(reco_for_user.iloc[0])
-
-
-def main():
-    model_user_knn = my_UserKnn(CosineRecommender(), list_pop_items_14d, 50)
-    model_user_knn.fit(df_inter)
-
-    # save model
-    with open("./models/user_knn/model_user_knn.dill", "wb") as f:
-        dill.dump(model_user_knn, f)
+    return list(final_reco)
 
 
 # For HW4********************************************************************
-
-
-if __name__ == "__main__":
-    main()
